move github summary feature to apps folder
